+++ src/reduxStore/slices/accountInfoSlice.ts
@@ -2,6 +2,7 @@ import { Address } from '@multiversx/sdk-core';
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 import { REHYDRATE } from 'redux-persist';
 import { ZERO } from 'constants/index';
+import { BatchTransactionsWSResponseType } from 'types';
 import { AccountType } from 'types/account.types';
 import { storage } from 'utils/storage';
 import { localStorageKeys } from 'utils/storage/local';
@@ -11,7 +12,6 @@ import {
   logoutAction,
   LoginActionPayloadType
 } from '../commonActions';
-import { BatchTransactionsWSResponseType } from 'types';
 
 export interface LedgerAccountType {
   index: number;