+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 - [Guardian screen pre-release](https://github.com/multiversx/mx-sdk-dapp/pull/739)]
+- [Fixed transactions data and signature parsing](https://github.com/multiversx/mx-sdk-dapp/pull/738)]
 
 ## [[v2.11.3]](https://github.com/multiversx/mx-sdk-dapp/pull/737)] - 2023-04-21
 - [Added support for decoding login token within `decodeNativeAuthToken](https://github.com/multiversx/mx-sdk-dapp/pull/736)]